# menu/openapi.snapshot.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "TAWROS API",
        "version": "1.0.0",
        "description": "API REST para frontend TAWROS, integrações e automações"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "paths": {
...
        "/api/v1/lookups/{resource}": {
            "get": {
                "summary": "Lookup simples por recurso",
                "tags": [
                    "Lookups"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
...
        "/api/v1/integrations/status": {
            "get": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/webhooks/{event}": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "event",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/{resource}/pull": {
            "get": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/{resource}/delta": {
            "get": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/{resource}/push": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/{resource}/upsert": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/import/{resource}": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/integrations/export/{resource}": {
            "get": {
                "tags": [
                    "Integrations"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "resource",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://192.168.15.26:3000"
        }
    ]
}